+++ packages/google_maps_flutter/google_maps_flutter/example/lib/place_advanced_marker.dart
@@ -247,10 +247,10 @@ class _PlaceAdvancedMarkerBodyState extends State<_PlaceAdvancedMarkerBody> {
 
   Future<void> _changeZIndex(MarkerId markerId) async {
     final AdvancedMarker marker = markers[markerId]!;
-    final double current = marker.zIndex;
+    final int current = marker.zIndexInt;
     setState(() {
       markers[markerId] = marker.copyWith(
-        zIndexParam: current == 12.0 ? 0.0 : current + 1.0,
+        zIndexIntParam: current == 12 ? 0 : current + 1,
       );
     });
   }