+++ packages/webview_flutter/webview_flutter_android/test/android_webview_cookie_manager_test.mocks.dart
@@ -106,6 +106,14 @@ class MockCookieManager extends _i1.Mock implements _i2.CookieManager {
           as _i5.Future<void>);
 
   @override
+  _i5.Future<String?> getCookies(String? url) =>
+      (super.noSuchMethod(
+            Invocation.method(#getCookies, [url]),
+            returnValue: _i5.Future<String?>.value(),
+          )
+          as _i5.Future<String?>);
+
+  @override
   _i2.CookieManager pigeon_copy() =>
       (super.noSuchMethod(
             Invocation.method(#pigeon_copy, []),